Disable org-pdftools
org-pdftools use a link instead of relative position to bind notes
and pdf file together.
It is more precious. However, since the location is a link,
org-roam will treat it as a hook and try jump to that file.
As a result, the link has no meaning anymore but inconvenient.

One of the solution is to disable org-roam mini mode when we are
working with pdf-view mode. But I do not know if this is working.
Because we are writing in org mode and org-roam is hooked with it.
And I do not feel the specific link is that useful to me.